Ugly copypaste for OutputCache
Even though the disk cache works (for all endpoints, meh) it doesn't set the
Cache-Control headers - I have no idea why. Maybe it's not intended to
and should be used in pair with the Response Cache Middleware, which sucks.
However there are other limits to the Middleware (i.e. when it doesn't work)
that don't make me happy.

So maybe the right way would be to skip all this middleware noncence
and implement the ImageCache Service as a layer between the controller
and the resizer (although I don't have the resizer service). It looks like
this would give me much more control over resized image caching
and would be easy-breezy